rewrite board
