+++ pyproject.toml
@@ -7,6 +7,7 @@ authors = ["Lucas Lee <replicat@users.noreply.github.com>"]
 [tool.poetry.dependencies]
 python = "^3.9"
 PuLP = "^2.6.0"
+tabulate = "^0.8.9"
 
 [tool.poetry.dev-dependencies]
 pre-commit = "^2.19.0"
