+++ src/board.py
@@ -1,140 +1,68 @@
 from __future__ import annotations
 
+import random
+
 import pulp as pl
+import tabulate
 
 from enums import Direction, Terrain
-from info import Info, RangeInfo
-
-# Currently only available for 5-players game
 
 
 class Board:
-    def __init__(self, info: list[Info] | None = None):
-        # preset variables
-        self.players: tuple = ("a", "b", "c", "d", "e", "public", "loot")
-        self.max_slot: int = 8 * 3
+    """Representation of a game board."""
 
-        # info available
-        self._info: list[Info] = info or []
+    def __init__(
+        self,
+        players: tuple,
+        terrains: tuple,
+        directions: tuple,
+    ):
+        """Constructor of `Board` class.
 
-        # lp problem
-        self._problem = pl.LpProblem(sense=pl.LpMaximize)
-        self._slots: dict[tuple, pl.LpVariable] = pl.LpVariable.dicts(
-            name="Slots",
-            indices=[(p, s) for p in self.players for s in range(self.max_slot)],
+        Args:
+            players (tuple): Tuple of players.
+            terrains (tuple): Tuple of types of terrain.
+            directions (tuple): Tuple of directions.
+        """
+        self._players: tuple = players
+        self._terrains: tuple = terrains
+        self._directions: tuple = directions
+
+        self.locations: dict[tuple, pl.LpVariable] = pl.LpVariable.dicts(
+            name="Player Locations",
+            indices=[(p, t, d) for p in self.players for t in self.terrains for d in self.directions],
             cat="Binary",
         )
+        for v in self.locations.values():
+            v.setInitialValue(random.randint(0, 1))
 
-        # init aggregate
-        self._aggregated: dict[tuple, float] = {}
-        self._count: int = 0
-        self._init_aggregate()
+    @property
+    def players(self) -> tuple:
+        return self._players
 
-    def _init_aggregate(self):
-        self._aggregated = {(p, s): 0.0 for p in self.players for s in range(self.max_slot)}
-        self._count = 0
+    @property
+    def terrains(self) -> tuple:
+        return self._terrains
 
-    def _get_slots(self, player: str, terrain: Terrain, start: Direction, end: Direction) -> list[pl.LpVariable]:
-        if terrain == Terrain.ALL:
-            return [
-                slot
-                for terrain in (Terrain.BEACH, Terrain.FOREST, Terrain.MOUNTAIN)
-                for slot in self._get_slots(player, terrain, start, end)
-            ]
-
-        # separate slots by terrain
-        terrain_slots = {
-            Terrain.BEACH: [self._slots[(player, slot)] for slot in range(0, 8)],
-            Terrain.FOREST: [self._slots[(player, slot)] for slot in range(8, 16)],
-            Terrain.MOUNTAIN: [self._slots[(player, slot)] for slot in range(16, 24)],
-        }
-
-        # translate direction to position
-        start_pos = start.value
-        end_pos = end.value
-
-        if end_pos <= start_pos:
-            return terrain_slots[terrain][start_pos:8] + terrain_slots[terrain][0:end_pos]
-        else:
-            return terrain_slots[terrain][start_pos:end_pos]
+    @property
+    def directions(self) -> tuple:
+        return self._directions
 
     def print(self):
-        if self._count <= 0:
-            if self._problem.status == pl.LpStatusNotSolved:
-                print("Not solved")
-            elif self._problem.status == pl.LpStatusInfeasible:
-                print("No solution found")
-            return
-
-        print("========================================")
-        for player in self.players:
-            print(player)
-            print("----------------------------------------")
-            for slot in range(self.max_slot):
-                print(f"{self._aggregated[(player, slot)] / self._count:.2f}", end=" ")
-                if slot in (7, 15, 23):
-                    print()
-            print("========================================")
-        print(f"Total Iterations: {self._count}")
-
-    def solve(self, max_iterations: int = 100):
-        self._init_aggregate()
-
-        # all slots must present once and only once
-        for slot in range(self.max_slot):
-            self._problem.addConstraint(pl.lpSum(self._slots[(player, slot)] for player in self.players) == 1)
-
-        # each real player have a total of 4 slots
-        for player in ("a", "b", "c", "d", "e"):
-            self._problem.addConstraint(pl.lpSum(self._slots[(player, slot)] for slot in range(self.max_slot)) == 4)
-
-        # public & loot each have a total of 2 slots
-        for player in ("public", "loot"):
-            self._problem.addConstraint(pl.lpSum(self._slots[(player, slot)] for slot in range(self.max_slot)) == 2)
-
-        # constraints from info
-        for info in self._info:
-            if isinstance(info, RangeInfo):
-                self._problem.addConstraint(
-                    pl.lpSum(self._get_slots(info.player, info.terrain, info.start, info.end)) == info.amount
-                )
-
-        for _ in range(max_iterations):
-            # solve once
-            self._problem.solve(pl.PULP_CBC_CMD(msg=0))
-            if self._problem.status != pl.LpStatusOptimal:
-                break
-
-            # add to aggregate
-            for player in self.players:
-                for slot in range(self.max_slot):
-                    self._aggregated[(player, slot)] += pl.value(self._slots[(player, slot)])
-            self._count += 1
-
-            # add constraint to prevent same solution
-            solution = [
-                self._slots[(player, slot)]
-                for player in self.players
-                for slot in range(self.max_slot)
-                if pl.value(self._slots[(player, slot)]) == 1
-            ]
-            self._problem.addConstraint(pl.lpSum(solution) <= self.max_slot - 1)
+        print("=" * 44)
+        for p in self.players:
+            print(p)
+            print("-" * 44)
+            table = [[t] + [f"{pl.value(self.locations[(p, t, d)]):f}" for d in self.directions] for t in self.terrains]
+            headers = [""] + list(self.directions)
+            print(tabulate.tabulate(table, headers, tablefmt="simple", floatfmt=".0f", numalign="center"))
+            print("=" * 44)
 
 
 if __name__ == "__main__":
-    info = [
-        RangeInfo(player="a", terrain=Terrain.MOUNTAIN, start=Direction.NORTH, end=Direction.SOUTH, amount=4),
-        RangeInfo(player="b", terrain=Terrain.FOREST, start=Direction.NORTH, end=Direction.SOUTH, amount=4),
-        RangeInfo(player="c", terrain=Terrain.ALL, start=Direction.NORTH, end=Direction.SOUTH, amount=4),
-        RangeInfo(player="d", terrain=Terrain.ALL, start=Direction.SOUTH, end=Direction.NORTH, amount=4),
-        RangeInfo(player="d", terrain=Terrain.FOREST, start=Direction.EAST, end=Direction.WEST, amount=2),
-        RangeInfo(player="d", terrain=Terrain.MOUNTAIN, start=Direction.EAST, end=Direction.WEST, amount=2),
-        RangeInfo(player="e", terrain=Terrain.ALL, start=Direction.SOUTH, end=Direction.NORTH, amount=4),
-        RangeInfo(player="e", terrain=Terrain.BEACH, start=Direction.SOUTH, end=Direction.SOUTH, amount=2),
-        RangeInfo(player="e", terrain=Terrain.FOREST, start=Direction.SOUTH, end=Direction.SOUTH, amount=0),
-        RangeInfo(player="public", terrain=Terrain.BEACH, start=Direction.SOUTH, end=Direction.SOUTHWEST, amount=1),
-        RangeInfo(player="public", terrain=Terrain.BEACH, start=Direction.WEST, end=Direction.NORTHWEST, amount=1),
-    ]
-    board = Board(info)
-    board.solve()
+    board = Board(
+        players=("A", "B", "C", "D", "E", "Public", "Loot"),
+        terrains=(Terrain.BEACH, Terrain.FOREST, Terrain.MOUNTAIN),
+        directions=Direction,
+    )
     board.print()
